% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGWAS-package.R
\docType{package}
\name{ggGWAS-package}
\alias{ggGWAS}
\alias{ggGWAS-package}
\title{ggGWAS: ggplot2 Extensions for Plotting GWAS Results}
\description{
Provides Manhattan plot and QQplot as a ggplot2 extension. 
    For high-dimensional GWAS data there is a hex version of the QQplot (which is more of a diagnostics tool).
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/sinarueeger/ggGWAS/issues}
}

}
\keyword{internal}
