% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlog10_trans.R
\name{log_sub_breaks_gggwas}
\alias{log_sub_breaks_gggwas}
\title{log sub breaks (copy)
exact copy from scales:::log_sub_breaks}
\usage{
log_sub_breaks_gggwas(rng, n = 5, base = 10)
}
\arguments{
\item{rng}{Range of values}

\item{n}{number of breaks}

\item{base}{log base}
}
\description{
log sub breaks (copy)
exact copy from scales:::log_sub_breaks
}
