% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlog10_trans.R
\name{mlog10_breaks}
\alias{mlog10_breaks}
\title{Breaks for -log10 axis}
\usage{
mlog10_breaks(n = 10)
}
\arguments{
\item{n}{number of breaks}
}
\value{
vector
}
\description{
Breaks for -log10 axis
}
\references{
From \url{https://stat.ethz.ch/pipermail/r-help/2014-April/373728.html} and \url{https://gist.github.com/wch/3250485}
}
