% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlog10_trans.R
\name{mlog10_trans}
\alias{mlog10_trans}
\title{Minus log-10 transformation}
\usage{
mlog10_trans()
}
\description{
Minus log-10 transformation
}
\examples{
require(ggplot2)
df <- data.frame(y = runif(1000))

ggplot(df, aes(sample = y)) +
  stat_qq(distribution = stats::qunif) +
  scale_y_continuous(trans = "mlog10") +
  scale_x_continuous(trans = "mlog10") +
  geom_abline(intercept = 0, slope = 1)
}
