% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlog10_trans.R
\name{mlog_breaks}
\alias{mlog_breaks}
\title{Minus log breaks (integer breaks on -log-trnsformed scales)}
\usage{
mlog_breaks(n = 5, base = 10)
}
\arguments{
\item{n}{desired number of breaks}

\item{base}{base of logarithm to use}
}
\value{
function
}
\description{
Minus log breaks (integer breaks on -log-trnsformed scales)
}
\examples{
require(scales)
mlog_breaks()(runif(1000))

}
\references{
From \url{https://stat.ethz.ch/pipermail/r-help/2014-April/373728.html} and \url{https://gist.github.com/wch/3250485}
}
