% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlog10_trans.R
\name{mlog_trans}
\alias{mlog_trans}
\title{Minus log transformation}
\usage{
mlog_trans(base = 10)
}
\arguments{
\item{base}{base of logarithm to use}
}
\description{
Minus log transformation
}
\examples{
require(ggplot2)
df <- data.frame(y = runif(1000))

ggplot(df, aes(sample = y)) +
  stat_qq(distribution = stats::qunif) +
  scale_y_continuous(trans = "mlog") +
  scale_x_continuous(trans = "mlog") +
  geom_abline(intercept = 0, slope = 1)
}
\references{
\url{https://stat.ethz.ch/pipermail/r-help/2014-April/373728.html and https://gist.github.com/wch/3250485}
}
