% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gwas_manhattan.R
\name{stat_gwas_manhattan}
\alias{stat_gwas_manhattan}
\alias{geom_gwas_manhattan}
\title{Manhattan plot}
\usage{
stat_gwas_manhattan()

geom_gwas_manhattan()
}
\value{
ggplot2 object
}
\description{
Manhattan plot for GWAS data.
}
\details{
The functions geom_gwas_manhattan and stat_gwas_manhattan are under construction. Use ggman::ggmanhattan() from https://github.com/mkanai/ggman in the meantime. \code{devtools::install_github('mkanai/ggman')}.
}
