% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gwas_qq_hex.R
\name{stat_gwas_qq_hex}
\alias{stat_gwas_qq_hex}
\alias{geom_gwas_qq_hex}
\title{Efficient Q-Q plot}
\usage{
stat_gwas_qq_hex(mapping = NULL, data = NULL, geom = "hex",
  position = "identity", na.rm = FALSE, bins = 30, binwidth = NULL,
  show.legend = NA, inherit.aes = TRUE, y.thresh = NULL,
  hex.function = hexBinSummarise, fill = "black", ...)

geom_gwas_qq_hex(mapping = NULL, data = NULL, geom = "hex",
  position = "identity", na.rm = FALSE, bins = 30, binwidth = NULL,
  show.legend = NA, inherit.aes = TRUE, y.thresh = NULL,
  hex.function = hexBinSummarise, fill = "black", ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{Override the default connection between \code{geom_hex} and
\code{stat_binhex.}}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{bins}{numeric vector giving number of bins in both vertical and
horizontal directions. Set to 30 by default.}

\item{binwidth}{Numeric vector giving bin width in both vertical and
horizontal directions. Overrides \code{bins} if both set.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{y.thresh}{Same scale as y (e.g. 0.05),
y <= y.thresh AFTER computing expected.}

\item{hex.function}{\code{ggGWAS:::hexBinSummarise} or \code{ggplot2:::hexBinSummarise}}

\item{fill}{color by which hexagons are filled, by default black.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Density-like QQ-plot with hexagons for high-dimensional data.
}
\details{
Code and documentation mostly from
\url{https://github.com/tidyverse/ggplot2/blob/master/R/stat-binhex.r}.
}
\examples{
require(ggplot2)

n.sample <- 1000
df <- data.frame(P = runif(n.sample), GWAS = sample(c("a", "b"),
  n.sample,
  replace = TRUE
))
theme_set(theme_bw())

(qp <- ggplot(df, aes(y = P)) +
  stat_gwas_qq_hex() +
  geom_abline(intercept = 0, slope = 1))

(qp <- ggplot(df, aes(y = P, group = GWAS, color = GWAS)) +
  stat_gwas_qq_hex() +
  geom_abline(intercept = 0, slope = 1))
}
\seealso{
\code{\link[ggplot2]{stat_bin_hex}}
Variables computed by \code{stat_gwas_qq_hex}:
\describe{
\item{y}{Observed P-value quantiles}
\item{x}{Expected/theoretical quantiles}
}
}
