% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_manhattan.R
\name{stat_manhattan}
\alias{stat_manhattan}
\title{Manhattan plot with ggplot2 features}
\usage{
stat_manhattan(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, y.thresh = NULL, ...)
}
\arguments{
\item{y.thresh}{a vector}

\item{...}{}
}
\description{
Manhattan plot with ggplot2 features
}
\examples{
dat <- qqman::gwasResults# \%>\% filter(P < 0.05)#, chr="CHR", bp="BP", snp="SNP", p="P" )


qp <- ggplot(dat \%>\% mutate(CHR2 = as.character(CHR))) +
 stat_manhattan(aes(pos = BP, y = -log10(P), chr = CHR)) +
 geom_hline(yintercept = 8) +
 ggtitle("sfsdfsdf")
print(qp)


}
