% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_qqplot.R
\name{stat_qqplot}
\alias{stat_qqplot}
\title{QQ-plot}
\usage{
stat_qqplot(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
QQ-plot
}
\examples{
n.sample <- 10000
df <- data.frame(P = runif(n.sample), GWAS = sample(c("a","b"), n.sample, replace = TRUE))

## default
qp <- ggplot(df, aes(observed = P)) +
  stat_qqplot() +
  geom_abline(intercept = 0, slope = 1)
print(qp)

## adding nice stuff
qp +
  theme(aspect.ratio=1) + ## square shaped
  expand_limits(x = -log10(max(df$P)), y = -log10(max(df$P))) + ## identical limits (meaning truely square)
  ggtitle("QQplot") + ## title
  xlab("Expected -log10(P)") + ## axis labels
  ylab("Observed -log10(P)")

## color
ggplot(df, aes(observed = P, color = GWAS)) +
  stat_qqplot() +
  geom_abline(intercept = 0, slope = 1)

## facet
ggplot(df, aes(observed = P)) +
  facet_wrap(~GWAS) +
  stat_qqplot() +
  geom_abline(intercept = 0, slope = 1)


## group
ggplot(df, aes(observed = P, group = GWAS)) +
  stat_qqplot() +
  geom_abline(intercept = 0, slope = 1)
}
