% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_qqunif_hex.R
\name{stat_qqunif_hex}
\alias{stat_qqunif_hex}
\title{QQ-plot}
\usage{
stat_qqunif_hex(mapping = NULL, data = NULL, geom = "hex",
  position = "identity", na.rm = FALSE, bins = 30, binwidth = NULL,
  show.legend = NA, inherit.aes = TRUE, observed.thresh = NULL, ...)
}
\arguments{
\item{binwidth}{}

\item{observed.thresh}{same scale as observed (e.g. 0.05), observed <= observed.thresh AFTER computing expected}
}
\description{
QQ-plot
}
\examples{
require(ggplot2)

n.sample <- 1000
df <- data.frame(P = runif(n.sample), GWAS = sample(c("a","b"), n.sample, replace = TRUE))


(qp <- ggplot(df, aes(y = P)) +
  stat_qqunif_hex() +
  geom_abline(intercept = 0, slope = 1))

(qp <- ggplot(df, aes(y = P, group = GWAS, color = GWAS)) +
  stat_qqunif_hex() +
  geom_abline(intercept = 0, slope = 1))



}
