% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gwas.R
\name{theme_gwas}
\alias{theme_gwas}
\title{Theme GWAS (just a random theme, nothing really gwas specific)}
\usage{
theme_gwas(base_size = 11, base_family = "")
}
\arguments{
\item{base_family}{}
}
\description{
Theme GWAS (just a random theme, nothing really gwas specific)
}
