% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gwas_manhattan.R, R/geom_gwas_qq.R,
%   R/geom_gwas_qq_hex.R, R/utilities.R
\docType{data}
\name{StatGwasManhattan}
\alias{StatGwasManhattan}
\alias{StatGwasQqplot}
\alias{StatGwasQqplotHex}
\alias{ggGWAS-ggproto}
\title{ggproto copy}
\description{
ggproto copy
}
\keyword{internal}
