% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggGWAS-package.R
\docType{package}
\name{ggGWAS-package}
\alias{ggGWAS}
\alias{ggGWAS-package}
\title{ggGWAS: ggplot2 Extensions for Plotting GWAS Summary Statistics}
\description{
Provides Manhattan plot and Q-Q plot as a ggplot2
    extension. For high-dimensional GWAS data there is a hex version of
    the Q-Q plot.
}
\author{
\strong{Maintainer}: Sina Rüeger \email{sina.rueeger@gmail.com}

}
\keyword{internal}
