% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlog10_trans.R
\name{mlog10_trans}
\alias{mlog10_trans}
\title{-log10 transformation}
\usage{
mlog10_trans()
}
\description{
-log10 transformation
}
\examples{
require(ggplot2)
df <- data.frame(y = runif(1000))

ggplot(df, aes(sample = y)) +
stat_qq(distribution = stats::qunif) +
scale_y_continuous(trans = mlog10_trans()) +
scale_x_continuous(trans = mlog10_trans()) +
geom_abline(intercept = 0, slope = 1)
}
\references{
\url{https://stat.ethz.ch/pipermail/r-help/2014-April/373728.html and https://gist.github.com/wch/3250485}
}
