% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gwas_manhattan.R
\name{stat_gwas_manhattan}
\alias{stat_gwas_manhattan}
\alias{geom_gwas_manhattan}
\title{Manhattan plot}
\usage{
stat_gwas_manhattan(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, y.thresh = NULL, chr.class = "numeric", ...)

geom_gwas_manhattan(mapping = NULL, data = NULL, geom = "point",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, y.thresh = NULL, chr.class = "numeric", ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}} or
\code{\link[=aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data.}

\item{geom}{\code{"point"} by default}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[=borders]{borders()}}.}

\item{y.thresh}{cutoff for y-axis, defined as a vector of length two.
If \code{c(K, NA)}, points with a y-value lower than K will be removed.
If \code{c(NA, K)} points with y-values larger than K will be removed.}

\item{chr.class}{Relevant for coloring of the points: what class the
chromosomes should be represented as. If "numeric", coloring will be
continuous. If "character" coloring will be discrete.}

\item{...}{Other arguments passed on to \code{\link[=layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\value{
sdfsdf?
}
\description{
Manhattan plot for GWAS data
}
\details{
See also \url{https://github.com/tidyverse/ggplot2/blob/master/R/stat-qq.r}
}
\examples{
require(ggplot2)
require(GWAS.utils) ## devtools::install_github("sinarueeger/GWAS.utils")
data("giant")
?giant
theme_set(theme_bw())

## default: for -log10(P), by default chr is numeric
qp <- ggplot(giant) +
  stat_gwas_manhattan(aes(pos = POS, y = -log10(P), chr = CHR)) +
  geom_hline(yintercept = 8) +
  ggtitle("GIANT summary statistics (by default CHR is numeric)")
print(qp)

## add nice color palette
pal <- wesanderson::wes_palette("Zissou1", 22, type = "continuous")
qp + scale_color_gradientn(colours = pal)

## chr factor
qp <- ggplot(giant) +
  stat_gwas_manhattan(aes(pos = POS, y = -log10(P), chr = CHR),
    chr.class = "character"
  ) +
  geom_hline(yintercept = 8) +
  ggtitle("GIANT summary statistics (CHR is now a character/factor)")
print(qp)
## adding a nice color palette
qp + scale_color_manual(values = wesanderson::wes_palette("Zissou1", 22,
  type = "continuous"
))

## turn all points black
qp <- ggplot(giant) +
  stat_gwas_manhattan(aes(pos = POS, y = -log10(P), chr = CHR),
    color = "black", alpha = I(0.4)
  ) +
  geom_hline(yintercept = 8) +
  ggtitle("GIANT summary statistics")
print(qp)

## set lower threshold
qp <- ggplot(data = giant) +
  stat_gwas_manhattan(aes(pos = POS, y = -log10(P), chr = CHR),
    y.thresh = c(2, NA)
  ) +
  geom_hline(yintercept = 8) +
  ggtitle("GIANT summary statistics")
print(qp)


## for effect sizes
qp <- ggplot(data = giant) +
  stat_gwas_manhattan(aes(pos = POS, y = BETA, chr = CHR)) +
  ggtitle("GIANT effect sizes")
print(qp)

## use rastr
qp <- ggplot(data = giant) +
  stat_gwas_manhattan(aes(pos = POS, y = -log10(P), chr = CHR),
    geom = ggrastr:::GeomPointRast
  ) +
  geom_hline(yintercept = 8) +
  ggtitle("GIANT summary statistics (rastr)")
print(qp)

## facetting

## generate two groups
giant <- giant \%>\%
  dplyr::mutate(gr = dplyr::case_when(
    BETA <= 0 ~ "Neg effect size",
    BETA > 0 ~ "Pos effect size"
  ))
qp <- ggplot(data = giant) +
  stat_gwas_manhattan(aes(pos = POS, y = BETA, chr = CHR)) +
  ggtitle("GIANT summary statistics") +
  facet_wrap(~gr)
print(qp)
}
